Make start_params_helpers work with unprocessed constraints.
